changes to_i to to_f
fixes issue that breaks floats eg. 1.5 hours becoming 1 hour

+++ lib/liquid_pluralize.rb
@@ -1,7 +1,7 @@
 module Pluralize
 
 	def pluralize(number, singular, plural = nil)
-		number = number.to_i
+		number = number.to_f
 		if number == 1
 			"#{number} #{singular}"
 		elsif plural.nil?
